correcion erores en finanzas y decimales del iva
.

+++ database/migrations/2022_10_08_233917_create_ivas_table.php
@@ -15,7 +15,7 @@ class CreateIvasTable extends Migration
     {
         Schema::create('ivas', function (Blueprint $table) {
             $table->id();
-            $table->integer('porcentaje');
+            $table->decimal('porcentaje', 12, 2);
             $table->string('descripcion');
             $table->string('usuario_edito');
             $table->timestamps();
